+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/graph/model/WhileItem.java
@@ -152,6 +152,20 @@ public class WhileItem extends LoopItem implements Block {
     }
 
     @Override
+    public List<BreakItem> getBreaks() {
+        List<BreakItem> ret = new ArrayList<>();
+        for (GraphTargetItem ti : commands) {
+            if (ti instanceof BreakItem) {
+                ret.add((BreakItem) ti);
+            }
+            if (ti instanceof Block) {
+                ret.addAll(((Block) ti).getBreaks());
+            }
+        }
+        return ret;
+    }
+    
+    @Override
     public boolean needsSemicolon() {
         return false;
     }