+++ libsrc/ffdec_lib/test/com/jpexs/decompiler/flash/ActionScript2Test.java
@@ -2515,4 +2515,31 @@ public class ActionScript2Test extends ActionScript2TestBase {
                 + "f(tst);\r\n"
         );
     }
+
+    @Test
+    public void frame93_Test() {
+        compareSrc(93, "trace(\"testInnerSwitchNoLabel\");\r\n"
+                + "var a;\r\n"
+                + "var b;\r\n"
+                + "switch(a)\r\n"
+                + "{\r\n"
+                + "case 4:\r\n"
+                + "trace(\"4\");\r\n"
+                + "break;\r\n"
+                + "case 7:\r\n"
+                + "trace(\"7\");\r\n"
+                + "break;\r\n"
+                + "default:\r\n"
+                + "switch(b)\r\n"
+                + "{\r\n"
+                + "case 1:\r\n"
+                + "trace(\"1\");\r\n"
+                + "break;\r\n"
+                + "case 2:\r\n"
+                + "trace(\"2\");\r\n"
+                + "}\r\n"
+                + "}\r\n"
+                + "trace(\"end\");\r\n"
+        );
+    }
 }